Enhances homepage with top player and invites
Improves the homepage by displaying the top player's information,
including their username and the game they are known for.

Also, it adds a visual indicator on the user's avatar in the header
if they have pending team invites.

These changes enhance user engagement and provide better discoverability
of key information on the homepage.